requirements.txt
tqdm
pytube3
unidecode
prettytable
requests